requirements.txt
django==2.1.15
django-redis
pymysql
requests
